refactor(tools): move normalizeMerchantName to leaf module
src/utils/merchant.ts has no project imports, so live/transactions.ts
no longer runtime-imports tools.ts. This breaks the only runtime edge
from tools/live/* back into tools.ts, keeping the import graph acyclic
when the registry index starts importing the live tool definitions.
tools.ts re-exports the function so its public surface is unchanged.
