+++ src/tools/tools.ts
@@ -56,6 +56,7 @@ import {
   getAccountsTool,
   getConnectionStatusTool,
 } from './registry/accounts-system.js';
+import { normalizeMerchantName } from '../utils/merchant.js';
 import type { LiveCopilotDatabase } from '../core/live-database.js';
 import type { GraphQLClient } from '../core/graphql/client.js';
 import { GraphQLError } from '../core/graphql/client.js';
@@ -321,70 +322,7 @@ function getCategoryIdOrDefault(categoryId: string | null | undefined): string {
   return categoryId || DEFAULT_CATEGORY_ID;
 }
 
-/**
- * Normalize merchant names for better aggregation.
- *
- * Handles variations like:
- * - "APPLE.COM-BILL" vs "APPLE.COM/BILL"
- * - "UBER" vs "UBER EATS"
- * - "AMAZON.COM*..." vs "AMAZON MKTPL*..." vs "AMAZON GROCE*..."
- */
-export function normalizeMerchantName(name: string): string {
-  let normalized = name.toUpperCase().trim();
-
-  // Remove common suffixes/prefixes
-  normalized = normalized
-    .replace(/[*#].*$/, '') // Remove everything after * or #
-    .replace(/\s+/g, ' ') // Normalize whitespace
-    .replace(/[.,/-]+/g, ' ') // Replace punctuation with spaces
-    .trim();
-
-  // Common merchant normalizations
-  const merchantMappings: Record<string, string> = {
-    'APPLE COM BILL': 'APPLE',
-    'APPLE COM': 'APPLE',
-    'AMAZON COM': 'AMAZON',
-    'AMAZON MKTPL': 'AMAZON',
-    'AMAZON GROCE': 'AMAZON GROCERY',
-    'AMZN MKTP': 'AMAZON',
-    AMZN: 'AMAZON',
-    'UBER EATS': 'UBER EATS',
-    'UBER TRIP': 'UBER',
-    'UBER BV': 'UBER',
-    LYFT: 'LYFT',
-    STARBUCKS: 'STARBUCKS',
-    DOORDASH: 'DOORDASH',
-    GRUBHUB: 'GRUBHUB',
-    'NETFLIX COM': 'NETFLIX',
-    NETFLIX: 'NETFLIX',
-    SPOTIFY: 'SPOTIFY',
-    HULU: 'HULU',
-    'DISNEY PLUS': 'DISNEY+',
-    DISNEYPLUS: 'DISNEY+',
-    'HBO MAX': 'HBO MAX',
-    WALMART: 'WALMART',
-    TARGET: 'TARGET',
-    COSTCO: 'COSTCO',
-    WHOLEFDS: 'WHOLE FOODS',
-    'WHOLE FOODS': 'WHOLE FOODS',
-    'TRADER JOE': 'TRADER JOES',
-  };
-
-  // Check for known mappings
-  for (const [pattern, replacement] of Object.entries(merchantMappings)) {
-    if (normalized.includes(pattern)) {
-      return replacement;
-    }
-  }
-
-  // Return first 3 words for long names
-  const words = normalized.split(' ').filter((w) => w.length > 0);
-  if (words.length > 3) {
-    return words.slice(0, 3).join(' ');
-  }
-
-  return normalized || name;
-}
+export { normalizeMerchantName };
 
 /**
  * A single investment holding enriched with security metadata and computed returns.
