+++ src/utils/merchant.ts
@@ -1,0 +1,72 @@
+/**
+ * Merchant name normalization.
+ *
+ * Kept in a leaf module (no project imports) so both the cache-mode tool
+ * handlers (`tools/tools.ts`) and the live tools (`tools/live/*`) can use it
+ * without creating a runtime import cycle through the tool registry.
+ */
+
+/**
+ * Normalize merchant names for better aggregation.
+ *
+ * Handles variations like:
+ * - "APPLE.COM-BILL" vs "APPLE.COM/BILL"
+ * - "UBER" vs "UBER EATS"
+ * - "AMAZON.COM*..." vs "AMAZON MKTPL*..." vs "AMAZON GROCE*..."
+ */
+export function normalizeMerchantName(name: string): string {
+  let normalized = name.toUpperCase().trim();
+
+  // Remove common suffixes/prefixes
+  normalized = normalized
+    .replace(/[*#].*$/, '') // Remove everything after * or #
+    .replace(/\s+/g, ' ') // Normalize whitespace
+    .replace(/[.,/-]+/g, ' ') // Replace punctuation with spaces
+    .trim();
+
+  // Common merchant normalizations
+  const merchantMappings: Record<string, string> = {
+    'APPLE COM BILL': 'APPLE',
+    'APPLE COM': 'APPLE',
+    'AMAZON COM': 'AMAZON',
+    'AMAZON MKTPL': 'AMAZON',
+    'AMAZON GROCE': 'AMAZON GROCERY',
+    'AMZN MKTP': 'AMAZON',
+    AMZN: 'AMAZON',
+    'UBER EATS': 'UBER EATS',
+    'UBER TRIP': 'UBER',
+    'UBER BV': 'UBER',
+    LYFT: 'LYFT',
+    STARBUCKS: 'STARBUCKS',
+    DOORDASH: 'DOORDASH',
+    GRUBHUB: 'GRUBHUB',
+    'NETFLIX COM': 'NETFLIX',
+    NETFLIX: 'NETFLIX',
+    SPOTIFY: 'SPOTIFY',
+    HULU: 'HULU',
+    'DISNEY PLUS': 'DISNEY+',
+    DISNEYPLUS: 'DISNEY+',
+    'HBO MAX': 'HBO MAX',
+    WALMART: 'WALMART',
+    TARGET: 'TARGET',
+    COSTCO: 'COSTCO',
+    WHOLEFDS: 'WHOLE FOODS',
+    'WHOLE FOODS': 'WHOLE FOODS',
+    'TRADER JOE': 'TRADER JOES',
+  };
+
+  // Check for known mappings
+  for (const [pattern, replacement] of Object.entries(merchantMappings)) {
+    if (normalized.includes(pattern)) {
+      return replacement;
+    }
+  }
+
+  // Return first 3 words for long names
+  const words = normalized.split(' ').filter((w) => w.length > 0);
+  if (words.length > 3) {
+    return words.slice(0, 3).join(' ');
+  }
+
+  return normalized || name;
+}
